整合了家辉的业务代码，修改了一些实体类和接口

+++ src/main/java/com/watering/domain/DTO/score/ScoreAddDTO.java
@@ -13,9 +13,9 @@ import lombok.Data;
 public class ScoreAddDTO {
     private Integer empid;
     //态度评分
-    private float attitude;
+    private Double attitude;
     //能力评分
-    private float ability;
+    private Double ability;
     //评价描述
     private String detail;
 }
